move RBData in RB routine file and change RBData.parameters type
