+++ src/qibocal/protocols/characterization/randomized_benchmarking/utils.py
@@ -183,6 +183,7 @@ def resample_p0(data, sample_size=100, homogeneous: bool = True):
 def samples_to_p0s(data, qubit):
     p0s = []
     qubit_keys = [key for key in list(data.keys()) if key[0] == qubit]
+    print("FFFFFFFFFFFFFFFFFFFFF", data, qubit_keys)
     for key in qubit_keys:
         p0s.append(1 - np.count_nonzero(np.array(data[key])) / len(data[key]))
     return p0s
